requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
langgraph==0.3.29
openai==1.72.0
python-dotenv==1.1.0
pydantic==2.11.3
pydantic-settings==2.2.1 